requirements.txt
django==3.1.2
pillow
djangorestframework
markdown
django-filter
